pyproject.toml
[project]
name = "mcp-dynamic-schema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp[openai]>=3.0.0b2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
